+++ accounts/models.py
@@ -1,0 +1,14 @@
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+
+class Position(models.Model):
+    name = models.CharField(max_length=100, unique=True)
+
+    def __str__(self):
+        return self.name
+
+
+class Worker(AbstractUser):
+    position = models.ForeignKey(Position, on_delete=models.SET_NULL, null=True, related_name="workers")
+    profile_picture = models.ImageField(upload_to="profile_pictures", null=True, blank=True)
